seperate preprocessing API and command line
