+++ onnxruntime/python/tools/quantization/__init__.py
@@ -2,3 +2,4 @@ from .calibrate import CalibraterBase, CalibrationDataReader, CalibrationMethod,
 from .qdq_quantizer import QDQQuantizer
 from .quant_utils import QuantFormat, QuantType, write_calibration_table
 from .quantize import QuantizationMode, quantize_dynamic, quantize_static
+from .shape_inference import quant_pre_process
